Refactor Navbar to remove type-based build logic
Simplify the Navbar component initialization by removing the dependency on navbar type for the build process. Updated corresponding imports and stories to reflect this change and renamed the utility function for clarity.
