+++ frontend-next-migration/src/widgets/Navbar/model/getNavbarBuildBySize.ts
